Refatoração da classe FactoryUsuario

+++ src/Classes/FactoryUsuario.java
@@ -6,7 +6,7 @@ import Modelo.Usuario;
 
 public class FactoryUsuario {
 
-    public Usuario createUsuario(int tipo, String nome, int matricula, String senha) {
+    public Usuario createUsuario(byte tipo, String nome, int matricula, String senha) {
         if (tipo != 1 || tipo != 2 ) {
             System.out.println("Tipo invalido, informe o valor 1 ou 2");
         }
